+++ lib/src/js_require.dart
@@ -31,15 +31,19 @@ class JSRequire {
   /// This defaults to a valid JS identifier based on [package].
   final String identifier;
 
-  /// Whether the dependency is loaded lazily or not. When [lazy] is true,
-  /// the dependency is exposed via a lazy getter.
+  /// Whether the dependency is loaded lazily.
+  ///
+  /// A lazy-loaded dependency only executes the `require()` function the first
+  /// time the module identifier is referenced, rather than eagerly executing it
+  /// when this package is loaded.
   ///
   /// This defaults to false.
   final bool lazy;
 
-  /// Whether the dependency is optional or not. When [optional] is true,
-  /// failures in requiring the dependency will return null instead of throwing
-  /// an error.
+  /// Whether the dependency is optional.
+  ///
+  /// An optional dependency's module identifier will be `null` if a load fails
+  /// rather than throwing an error.
   ///
   /// This defaults to false.
   final bool optional;